+++ src/settings.ts
@@ -1,12 +1,14 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import { App, PluginSettingTab, Setting } from 'obsidian';
 import SteemitPlugin from './main';
-import { SteemitPluginSettings } from './types';
+import { RewardType, SteemitPluginSettings } from './types';
 
 export const DEFAULT_SETTINGS: SteemitPluginSettings = {
   category: '',
   username: '',
   password: '',
   appName: '',
+  rewardType: RewardType.DEFAULT,
 };
 
 export class SteemitSettingTab extends PluginSettingTab {
@@ -17,6 +19,13 @@ export class SteemitSettingTab extends PluginSettingTab {
     super(app, plugin);
     this.plugin = plugin;
     this.defaultAppName = `${this.plugin.manifest.id}/${this.plugin.manifest.version}`;
+  }
+
+  async saveSettings(name: keyof SteemitPluginSettings, value: string) {
+    if (this.plugin.settings) {
+      (this.plugin.settings as any)[name] = value;
+      await this.plugin.saveSettings();
+    }
   }
 
   display(): void {
@@ -31,27 +40,24 @@ export class SteemitSettingTab extends PluginSettingTab {
       .setDesc('Enter your Steemit username')
       .addText(text => {
         text
-          .setPlaceholder('anpigon')
+          .setPlaceholder('Your username')
           .setValue(this.plugin.settings?.username ?? '')
           .onChange(async value => {
-            this.plugin.settings!.username = value;
-            await this.plugin.saveSettings();
+            this.saveSettings('username', value);
           });
       });
 
     new Setting(containerEl)
       .setName('Password')
-      .setDesc(
-        'Enter your Steemit password (your Steemit privateKey for post).',
-      )
+      .setDesc('Enter your Steemit password (your Steemit privateKey for post).')
       .addText(text => {
         text
           .setPlaceholder('Your password')
           .setValue(this.plugin.settings?.password ?? '')
           .onChange(async value => {
-            this.plugin.settings!.password = value;
-            await this.plugin.saveSettings();
+            this.saveSettings('password', value);
           });
+        text.inputEl.type = 'password';
       });
 
     new Setting(containerEl)
@@ -59,13 +65,20 @@ export class SteemitSettingTab extends PluginSettingTab {
       .setDesc('Enter the category you want to post.')
       .addText(text => {
         text
-          .setPlaceholder('hive-101145')
+          .setPlaceholder('ex. hive-101145')
           .setValue(this.plugin.settings?.category ?? '')
           .onChange(async value => {
-            this.plugin.settings!.category = value;
-            await this.plugin.saveSettings();
+            this.saveSettings('category', value);
           });
       });
+
+    new Setting(containerEl).setName('Default Rewards').addDropdown(cb => {
+      cb.addOption('100%', 'Power Up 100%');
+      cb.addOption('50%', 'Default (50% / 50%)');
+      cb.addOption('0%', 'Decline Payout');
+      cb.setValue('50%');
+      cb.onChange(value => this.saveSettings('rewardType', value))
+    });
 
     new Setting(containerEl)
       .setName('Metadata AppName (options)')
@@ -75,8 +88,7 @@ export class SteemitSettingTab extends PluginSettingTab {
           .setPlaceholder(this.defaultAppName)
           .setValue(this.plugin.settings?.appName || this.defaultAppName)
           .onChange(async value => {
-            this.plugin.settings!.appName = value;
-            await this.plugin.saveSettings();
+            this.saveSettings('appName', value);
           });
       });
   }